+++ packages/python/m/ci/start_release.py
@@ -145,12 +145,6 @@ def after_checkout(branch_checkout: str, stashed: bool) -> OneOf[Issue, None]:
     return Good(None)
 
 
-def _branch_checkout(branch: str) -> OneOf[Issue, str]:
-    return git.checkout_branch(branch).flat_map_bad(
-        lambda err: issue('git checkout failure', cause=err),
-    )
-
-
 def _get_commits(gh_ver: str) -> OneOf[Issue, list[str] | None]:
     either = git.get_commits(gh_ver)
     if is_bad(either):
@@ -182,7 +176,7 @@ def start_release(gh_token: str, hotfix: bool = False) -> OneOf[Issue, None]:
         for commits in _get_commits(gh_ver)
         for _ in verify_release(commits, hotfix=hotfix)
         for new_ver in io.prompt_next_version(gh_ver, release_type)
-        for branch_checkout in _branch_checkout(f'{release_type}/{new_ver}')
+        for branch_checkout in git.checkout_branch(f'{release_type}/{new_ver}')
         for _ in after_checkout(branch_checkout, stashed_changes)
         for _ in release_setup('m', config, new_ver)
     ])