+++ packages/python/m/git.py
@@ -1,4 +1,4 @@
-from m.core import Good, Issue, OneOf, one_of, subprocess
+from m.core import Good, Issue, OneOf, issue, one_of, subprocess
 
 
 def get_branch() -> OneOf[Issue, str]:
@@ -37,7 +37,9 @@ def checkout_branch(branch: str) -> OneOf[Issue, str]:
     Returns:
         A `OneOf` containing an `Issue` of the git response.
     """
-    return subprocess.eval_cmd(f'git checkout -b {branch}')
+    return subprocess.eval_cmd(f'git checkout -b {branch}').flat_map_bad(
+        lambda err: issue('git checkout failure', cause=err),
+    )
 
 
 def get_first_commit_sha() -> OneOf[Issue, str]: