+++ packages/python/m/github/cli.py
@@ -4,8 +4,8 @@ from ..core import one_of
 from ..core.fp import Good, OneOf
 from ..core.issue import Issue
 from ..core.json import get
-from .api import graphql
 from .ci import create_ci_query
+from .graphql.api import graphql
 
 
 def get_pr_info(